{{cookiecutter.app_name}}/requirements/base.txt
# Core
pytz==2019.3
Django==3.0.1
Pillow==6.2.1

# Configuration manager 
django-configurations==2.2
python-decouple==3.3

# For the persistence stores
psycopg2-binary==2.8.3
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0

# Rest apis
djangorestframework==3.11.0
Markdown==3.1.1
django-filter==2.2.0
django-cors-headers==3.2.0
django-extra-fields==2.0.2
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
{% if cookiecutter.use_celery == 'y' %}

# Celery
celery==4.4.0
django-celery-beat==1.5.0
{% endif %}
